requirements.txt
# WFGameAI测试框架依赖

# 核心依赖
airtest>=1.3.0
adbutils>=2.8.0
opencv-python>=4.6.0
numpy>=1.20.0
ultralytics>=8.0.0
torch>=2.0.0
pillow>=9.0.0
roboflow>=1.1.0
supervision>=0.25.0

# UI测试相关
pyautogui>=0.9.50
mss>=6.0.0
pyobjc>=7.0.0; platform_system=="Darwin"  # 仅MacOS需要

# 工具库
jinja2>=3.0.0
matplotlib>=3.0.0
pyyaml>=6.0.0

# 网络和通信
requests>=2.24.0
websocket-client>=0.57.0

# 辅助工具
python-dateutil>=2.8.0
tqdm>=4.50.0
psutil>=5.7.0

# 日志和调试
logzero>=1.5.0
rich>=9.0.0

# 特殊设备支持
tidevice>=0.10.0

# AI引擎依赖
torchvision>=0.15.0

# 设备管理依赖
adbutils>=0.13.0
airtest>=1.2.5

# Web框架依赖
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.2
drf-yasg>=1.21.7
celery>=5.3.1
redis>=4.5.5
channels>=4.0.0

# 工具依赖
python-dotenv>=1.0.0
Jinja2>=3.1.2
PyYAML>=6.0
requests>=2.30.0
tqdm>=4.65.0

# 开发工具
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.3.1 